backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.5.3
pydantic-settings==2.1.0
celery==5.3.6
redis==5.0.1
httpx==0.26.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai==1.10.0
flower==2.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
google-auth==2.28.0
email-validator==2.1.0
